+++ src/core/core-varlink.h
@@ -8,7 +8,7 @@ void manager_varlink_done(Manager *m);
 
 /* Creates a new VarlinkServer and binds methods. Does not set up sockets or attach events.
  * Used for manager serialize/deserialize. */
-int manager_setup_varlink_server(Manager *m, VarlinkServer **ret_s);
+int manager_setup_varlink_server(Manager *m);
 
 /* The manager is expected to send an update to systemd-oomd if one of the following occurs:
  * - The value of ManagedOOM*= properties change
